COWER-46: Split solution
