+++ app/backend/Cower.Web/AuthorizeCheckOperationFilter.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
-namespace Cower.Presentation;
+namespace Cower.Web;
 
 public class AuthorizeCheckOperationFilter : IOperationFilter
 {
