+++ app/backend/Cower.Web/Models/RegisterResponseDTO.cs
@@ -1,4 +1,4 @@
-namespace Cower.Presentation.Models;
+namespace Cower.Web.Models;
 
 public record RegisterResponseDTO(
     UserResponseDTO User,
