--- app/backend/Cower/Cower.http
@@ -1,6 +1,0 @@
-@Cower_HostAddress = http://localhost:5233
-
-GET {{Cower_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
